# docs/swagger.yaml
basePath: /api
definitions:
...
  ds.User:
    properties:
      id:
        type: integer
      role:
        $ref: '#/definitions/role.Role'
      username:
        type: string
    type: object
...
  handler.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      role:
        type: string
      token_type:
        type: string
      user_id:
        type: integer
    type: object
  handler.RegisterRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handler.RegisterResponse:
    properties:
      message:
        type: string
      user_id:
        type: integer
    type: object
  handler.RemoveFromRequestRequest:
    properties:
      lamp_id:
        type: integer
      request_id:
        type: integer
    required:
    - lamp_id
    - request_id
    type: object
...
  handler.UpdateProfileRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handler.UpdateRequestLampRequest:
    properties:
      area_m2:
        type: number
      lamp_id:
        type: integer
      number:
        type: integer
      request_id:
        type: integer
    required:
    - lamp_id
    - request_id
    type: object
...
  role.Role:
    enum:
    - 0
    - 1
    type: integer
    x-enum-comments:
      Moderator: "1"
      User: "0"
    x-enum-descriptions:
    - "0"
    - "1"
    x-enum-varnames:
    - User
    - Moderator
host: localhost:8001
info:
  contact:
    email: support@example.com
    name: API Support
    url: https://github.com/yourusername/dia-backend
  description: REST API для веб-приложения расчета освещения комнат
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Let There Be Light API
  version: "1.0"
paths:
...
  /light-request-lamps:
    delete:
      consumes:
      - application/json
      description: Remove a lamp from a draft light request
      parameters:
      - description: Remove lamp data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RemoveFromRequestRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Remove lamp from request
      tags:
      - light-request-lamps
    put:
      consumes:
      - application/json
      description: Update area and/or number of lamps in a draft request
      parameters:
      - description: Update lamp data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateRequestLampRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update lamp in request
      tags:
      - light-request-lamps
...
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: User credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.LoginResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - users
  /users/logout:
    post:
      consumes:
      - application/json
      description: Invalidate JWT token by adding it to blacklist
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - users
  /users/profile:
    get:
      consumes:
      - application/json
      description: Get current authenticated user's profile information
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ds.User'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user profile
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update current authenticated user's profile (username and/or password)
      parameters:
      - description: Profile update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update user profile
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with username and password
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handler.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"